# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alloy Works",
    "version": "1"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-aw_session"
      }
    }
  },
  "paths": {
...
    "/v1/sign-in/google": {
      "get": {
        "operationId": "startGoogleSignIn",
        "summary": "Begin signing in with a Google account, by way of the one sign-in address",
        "security": [],
        "responses": {
          "302": {
            "description": "On to Google",
            "headers": {
              "Location": {
                "description": "Where to go next",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "This environment does not permit signing in this way",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every error takes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sign-in/google/callback": {
      "get": {
        "operationId": "finishGoogleSignIn",
        "summary": "Where Google returns, at the sign-in address only; hands the sign-in to its environment",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Admitted, and on to the environment that asked, with a one-time code",
            "headers": {
              "Location": {
                "description": "Where to go next",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "The sign-in could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "This account is not invited to that environment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "This is not the sign-in address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every error takes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sign-in/google/complete": {
      "get": {
        "operationId": "completeGoogleSignIn",
        "summary": "Redeems the one-time code from the sign-in address, and signs in",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Signed in, and on to the application",
            "headers": {
              "Location": {
                "description": "Where to go next",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "The sign-in could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every error takes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sign-in/organisation": {
      "get": {
        "operationId": "startOrganisationSignIn",
        "summary": "Begin signing in with the organisation's identity provider",
        "security": [],
        "responses": {
          "302": {
            "description": "On to the identity provider",
            "headers": {
              "Location": {
                "description": "Where to go next",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "This environment does not permit signing in this way",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every error takes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable and machine-readable: branch on this, never on the message"
                    },
                    "message": {
                      "type": "string",
                      "description": "For people. It may change between releases"
                    },
                    "rule": {
                      "description": "The requirement or rule that refused the request, where one did",
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Quote this when reporting a problem"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "traceId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}